Added sample API test, and updated project structure
