+++ pyproject.toml
@@ -1,22 +1,22 @@
 [tool.poetry]
-name = "playwright-pytest"
-version = "0.1.0"
-description = ""
-authors = ["Viacheslav <vacheslavuslistyj@gmail.com>"]
-readme = "README.md"
+name = 'playwright-pytest'
+version = '0.1.0'
+description = ''
+authors = ['Viacheslav <akaslavich@gmail.com>']
+readme = 'README.md'
 
 [tool.poetry.dependencies]
-python = "^3.12"
-pytest-playwright = "^0.4.3"
-ruff = "^0.1.8"
-pre-commit = "^3.6.0"
-pytest-xdist = "^3.5.0"
-allure-pytest = "^2.13.5"
+python = '^3.12'
+pytest-playwright = '^0.4.3'
+ruff = '^0.1.8'
+pre-commit = '^3.6.0'
+pytest-xdist = '^3.5.0'
+allure-pytest = '^2.13.5'
 
 
 [build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+requires = ['poetry-core']
+build-backend = 'poetry.core.masonry.api'
 
 
 
@@ -27,30 +27,30 @@ respect-gitignore = true
 # Exclude files or folder
 exclude = [
     # Default values
-    ".bzr",
-    ".direnv",
-    ".eggs",
-    ".git",
-    ".git-rewrite",
-    ".hg",
-    ".mypy_cache",
-    ".nox",
-    ".pants.d",
-    ".pytype",
-    ".ruff_cache",
-    ".svn",
-    ".tox",
-    "__pypackages__",
-    "_build",
-    "buck-out",
-    "build",
-    "dist",
-    "node_modules",
+    '.bzr',
+    '.direnv',
+    '.eggs',
+    '.git',
+    '.git-rewrite',
+    '.hg',
+    '.mypy_cache',
+    '.nox',
+    '.pants.d',
+    '.pytype',
+    '.ruff_cache',
+    '.svn',
+    '.tox',
+    '__pypackages__',
+    '_build',
+    'buck-out',
+    'build',
+    'dist',
+    'node_modules',
     # Virtualenvs
-    ".venv",
-    "venv",
+    '.venv',
+    'venv',
     # Folders with tests
-    "test",
+    'test',
 ]
 
 # Same as Black.
@@ -58,7 +58,7 @@ line-length = 120
 indent-width = 4
 
 # Assume Python 3.12
-target-version = "py312"
+target-version = 'py312'
 
 
 
@@ -67,51 +67,52 @@ target-version = "py312"
 # Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
 # McCabe complexity (`C901`) by default.
 select = [
-    "E", # pycodestyle
-    "F", # Pyflakes
-    "UP", # pyupgrade
-    "B", # flake8-bugbear
-    "SIM", # flake8-simplify
-    #    "I",  # isort
-    "YTT", # flake8-2020
-    "ASYNC", # flake8-async
-    "S", # flake8-nandit
-    "ICN", # flake8-import-conventions
-    "PT", # flake8-pytest-style
-    "ARG", # flake8-unused-arguments
-    "PL", # Pylint
-    "FLY", # flynt
-    "RUF", # Ruff-specific rules
+    'E', # pycodestyle
+    'F', # Pyflakes
+    'UP', # pyupgrade
+    'B', # flake8-bugbear
+    'SIM', # flake8-simplify
+    #    'I',  # isort
+    'YTT', # flake8-2020
+    'ASYNC', # flake8-async
+    'S', # flake8-nandit
+    'ICN', # flake8-import-conventions
+    'PT', # flake8-pytest-style
+    'ARG', # flake8-unused-arguments
+    'PL', # Pylint
+    'FLY', # flynt
+    'RUF', # Ruff-specific rules
 ]
 ignore = [
-    "PLW2901",
-    "S101",
-    "D300",
+    'D300',
+    'PLW2901',
+    'S101',
+    'UP035',
 ]
 # Allow fix for all enabled rules (when `--fix`) is provided.
-fixable = ["ALL"]
+fixable = ['ALL']
 unfixable = []
 # Allow unused variables when underscore-prefixed.
-dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
+dummy-variable-rgx = '^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$'
 
 
 [tool.ruff.lint.per-file-ignores]
-"__init__.py" = ["E402", "F401"]
+'__init__.py' = ['E402', 'F401']
 
 
 [tool.ruff.lint.flake8-quotes]
-docstring-quotes = "single"
+docstring-quotes = 'single'
 
 
 [tool.ruff.format]
 # Use single quotes for non-triple-quoted strings.
-quote-style = "single"
+quote-style = 'single'
 # Like Black, indent with spaces, rather than tabs.
-indent-style = "space"
+indent-style = 'space'
 # Like Black, respect magic trailing commas.
 skip-magic-trailing-comma = true
 # Like Black, automatically detect the appropriate line ending.
-line-ending = "auto"
+line-ending = 'auto'
 # Enable auto-formatting of code examples in docstrings. Markdown,
 # reStructuredText code/literal blocks and doctests are all supported.
 #
@@ -123,4 +124,4 @@ docstring-code-format = true
 #
 # This only has an effect when the `docstring-code-format` setting is
 # enabled.
-docstring-code-line-length = "dynamic"
+docstring-code-line-length = 'dynamic'
