+++ pytest.manual.ini
@@ -4,9 +4,3 @@ addopts = --alluredir allure-results
           --screenshot off
           --strict-markers
           --tracing off
-
-markers =
-    ui: marks tests that requires webdriver and tests UI flow. (select with '-m "ui"')
-    api: marks tests tests only API flow, without UI. (select with '-m "api"')
-    slow: marks tests as slow (deselect with '-m "not slow"')
-    internal: tests for logic verification of internal test automation modules.
